Fixed it!
Jeez that was harder than it should have been. Arduino was inserting
it's automatic prototypes in a commented out (by the preprocessor)
chunk of code. This was in addition to Arduino just failing to figure
out that the library declaration (AccelStepper) embeded deep in the
code was supposed to be imported from the Arduino Libraries folder!

This is what happens when invisible white magic gets bored with helping
innocent coders...